+++ tools/Manual/gmm_d.cpp
@@ -173,7 +173,7 @@ void gmm_objective_d(int d, int k, int n,
 
 #include "Eigen/Dense"
 
-#include "../../src/cpp/modules/manualEigen/gmm_eigen.h"
+#include "../../src/cpp/shared/gmm_eigen.h"
 
 using Eigen::Map;
 using Eigen::VectorXd;
